sorting on cluster table
