+++ report-viewer/src/components/ComparisonsTable.vue
@@ -3,135 +3,151 @@
 -->
 <template>
   <div class="flex flex-col">
-    <div class="font-bold">
-      <!-- Header -->
-      <div class="tableRow">
-        <div class="tableCellNumber"></div>
-        <div class="tableCellName items-center">Submissions in Comparison</div>
-        <div class="tableCellSimilarity !flex-col">
-          <div>Similarity</div>
-          <div class="flex w-full flex-row">
-            <ToolTipComponent class="flex-1" :direction="displayClusters ? 'top' : 'left'">
-              <template #default>
-                <p class="w-full text-center">{{ metricToolTips[MetricType.AVERAGE].shortName }}</p>
-              </template>
-              <template #tooltip>
-                <p class="whitespace-pre text-sm">
-                  {{ metricToolTips[MetricType.AVERAGE].tooltip }}
-                </p>
-              </template>
-            </ToolTipComponent>
+    <ComparisonTableFilter
+      v-model:search-string="searchString"
+      :enable-cluster-sorting="clusters != undefined"
+      :header="header"
+    />
 
-            <ToolTipComponent class="flex-1" :direction="displayClusters ? 'top' : 'left'">
-              <template #default>
-                <p class="w-full text-center">{{ metricToolTips[MetricType.MAXIMUM].shortName }}</p>
-              </template>
-              <template #tooltip>
-                <p class="whitespace-pre text-sm">
-                  {{ metricToolTips[MetricType.MAXIMUM].tooltip }}
-                </p>
-              </template>
-            </ToolTipComponent>
-          </div>
-        </div>
-        <div class="tableCellCluster items-center" v-if="displayClusters">Cluster</div>
-      </div>
-    </div>
+    <div class="flex flex-col overflow-hidden">
+      <div class="font-bold">
+        <!-- Header -->
+        <div class="tableRow">
+          <div class="tableCellNumber"></div>
+          <div class="tableCellName items-center">Submissions in Comparison</div>
+          <div class="tableCellSimilarity !flex-col">
+            <div>Similarity</div>
+            <div class="flex w-full flex-row">
+              <ToolTipComponent class="flex-1" :direction="displayClusters ? 'top' : 'left'">
+                <template #default>
+                  <p class="w-full text-center">
+                    {{ metricToolTips[MetricType.AVERAGE].shortName }}
+                  </p>
+                </template>
+                <template #tooltip>
+                  <p class="whitespace-pre text-sm">
+                    {{ metricToolTips[MetricType.AVERAGE].tooltip }}
+                  </p>
+                </template>
+              </ToolTipComponent>
 
-    <!-- Body -->
-    <div class="flex flex-grow flex-col overflow-hidden">
-      <DynamicScroller v-if="topComparisons.length > 0" :items="comparisonList" :min-item-size="48">
-        <template #default="{ item, index, active }">
-          <DynamicScrollerItem
-            :item="item"
-            :active="active"
-            :size-dependencies="[
-              item.firstSubmissionId,
-              item.secondSubmissionId,
-              store().isAnonymous(item.firstSubmissionId),
-              store().isAnonymous(item.secondSubmissionId)
-            ]"
-            :data-index="index"
-          >
-            <!-- Row -->
-            <div
-              class="tableRow"
-              :class="{
-                'bg-container-secondary-light dark:bg-container-secondary-dark': item.id % 2 == 1
-              }"
-            >
-              <div
-                @click="
-                  router.push({
-                    name: 'ComparisonView',
-                    params: { firstId: item.firstSubmissionId, secondId: item.secondSubmissionId }
-                  })
-                "
-                class="flex flex-grow cursor-pointer flex-row"
-              >
-                <!-- Index in sorted list -->
-                <div class="tableCellNumber">
-                  <div class="w-full text-center">{{ item.sortingPlace + 1 }}</div>
-                </div>
-
-                <!-- Names -->
-                <div class="tableCellName">
-                  <NameElement :id="item.firstSubmissionId" class="h-full w-1/2 px-2" />
-                  <NameElement :id="item.secondSubmissionId" class="h-full w-1/2 px-2" />
-                </div>
-
-                <!-- Similarities -->
-                <div class="tableCellSimilarity">
-                  <div class="w-1/2">
-                    {{ (item.similarities[MetricType.AVERAGE] * 100).toFixed(2) }}%
-                  </div>
-                  <div class="w-1/2">
-                    {{ (item.similarities[MetricType.MAXIMUM] * 100).toFixed(2) }}%
-                  </div>
-                </div>
-              </div>
-
-              <!-- Clusters -->
-              <div class="tableCellCluster flex !flex-col items-center" v-if="displayClusters">
-                <RouterLink
-                  v-for="index of getClusterIndexesFor(
-                    item.firstSubmissionId,
-                    item.secondSubmissionId
-                  )"
-                  v-bind:key="index"
-                  :to="{
-                    name: 'ClusterView',
-                    params: { clusterIndex: index }
-                  }"
-                  class="tect-center flex w-full justify-center"
-                >
-                  <ToolTipComponent class="w-fit" direction="left">
-                    <template #default>
-                      {{ clusters?.[index].members?.length }}
-                      <FontAwesomeIcon
-                        :icon="['fas', 'user-group']"
-                        :style="{ color: clusterIconColors[index] }"
-                      />
-                      {{ ((clusters?.[index].averageSimilarity as number) * 100).toFixed(2) }}%
-                    </template>
-                    <template #tooltip>
-                      <p class="whitespace-nowrap text-sm">
-                        {{ clusters?.[index].members?.length }} submissions in cluster with average
-                        similarity of
-                        {{ ((clusters?.[index].averageSimilarity as number) * 100).toFixed(2) }}%
-                      </p>
-                    </template>
-                  </ToolTipComponent>
-                </RouterLink>
-              </div>
+              <ToolTipComponent class="flex-1" :direction="displayClusters ? 'top' : 'left'">
+                <template #default>
+                  <p class="w-full text-center">
+                    {{ metricToolTips[MetricType.MAXIMUM].shortName }}
+                  </p>
+                </template>
+                <template #tooltip>
+                  <p class="whitespace-pre text-sm">
+                    {{ metricToolTips[MetricType.MAXIMUM].tooltip }}
+                  </p>
+                </template>
+              </ToolTipComponent>
             </div>
-          </DynamicScrollerItem>
-        </template>
+          </div>
+          <div class="tableCellCluster items-center" v-if="displayClusters">Cluster</div>
+        </div>
+      </div>
 
-        <template #after>
-          <slot name="footer"></slot>
-        </template>
-      </DynamicScroller>
+      <!-- Body -->
+      <div class="flex flex-grow flex-col overflow-hidden">
+        <DynamicScroller
+          v-if="topComparisons.length > 0"
+          :items="displayedComparisons"
+          :min-item-size="48"
+        >
+          <template #default="{ item, index, active }">
+            <DynamicScrollerItem
+              :item="item"
+              :active="active"
+              :size-dependencies="[
+                item.firstSubmissionId,
+                item.secondSubmissionId,
+                store().isAnonymous(item.firstSubmissionId),
+                store().isAnonymous(item.secondSubmissionId)
+              ]"
+              :data-index="index"
+            >
+              <!-- Row -->
+              <div
+                class="tableRow"
+                :class="{
+                  'bg-container-secondary-light dark:bg-container-secondary-dark': item.id % 2 == 1
+                }"
+              >
+                <div
+                  @click="
+                    router.push({
+                      name: 'ComparisonView',
+                      params: { firstId: item.firstSubmissionId, secondId: item.secondSubmissionId }
+                    })
+                  "
+                  class="flex flex-grow cursor-pointer flex-row"
+                >
+                  <!-- Index in sorted list -->
+                  <div class="tableCellNumber">
+                    <div class="w-full text-center">{{ item.sortingPlace + 1 }}</div>
+                  </div>
+
+                  <!-- Names -->
+                  <div class="tableCellName">
+                    <NameElement :id="item.firstSubmissionId" class="h-full w-1/2 px-2" />
+                    <NameElement :id="item.secondSubmissionId" class="h-full w-1/2 px-2" />
+                  </div>
+
+                  <!-- Similarities -->
+                  <div class="tableCellSimilarity">
+                    <div class="w-1/2">
+                      {{ (item.similarities[MetricType.AVERAGE] * 100).toFixed(2) }}%
+                    </div>
+                    <div class="w-1/2">
+                      {{ (item.similarities[MetricType.MAXIMUM] * 100).toFixed(2) }}%
+                    </div>
+                  </div>
+                </div>
+
+                <!-- Clusters -->
+                <div class="tableCellCluster flex !flex-col items-center" v-if="displayClusters">
+                  <RouterLink
+                    v-for="index of getClusterIndexesFor(
+                      item.firstSubmissionId,
+                      item.secondSubmissionId
+                    )"
+                    v-bind:key="index"
+                    :to="{
+                      name: 'ClusterView',
+                      params: { clusterIndex: index }
+                    }"
+                    class="flex w-full justify-center text-center"
+                  >
+                    <ToolTipComponent class="w-fit" direction="left">
+                      <template #default>
+                        {{ clusters?.[index].members?.length }}
+                        <FontAwesomeIcon
+                          :icon="['fas', 'user-group']"
+                          :style="{ color: clusterIconColors[index] }"
+                        />
+                        {{ ((clusters?.[index].averageSimilarity as number) * 100).toFixed(2) }}%
+                      </template>
+                      <template #tooltip>
+                        <p class="whitespace-nowrap text-sm">
+                          {{ clusters?.[index].members?.length }} submissions in cluster with
+                          average similarity of
+                          {{ ((clusters?.[index].averageSimilarity as number) * 100).toFixed(2) }}%
+                        </p>
+                      </template>
+                    </ToolTipComponent>
+                  </RouterLink>
+                </div>
+              </div>
+            </DynamicScrollerItem>
+          </template>
+
+          <template #after>
+            <slot name="footer"></slot>
+          </template>
+        </DynamicScroller>
+      </div>
     </div>
   </div>
 </template>
@@ -139,7 +155,7 @@
 <script setup lang="ts">
 import type { Cluster } from '@/model/Cluster'
 import type { ComparisonListElement } from '@/model/ComparisonListElement'
-import { toRef } from 'vue'
+import { computed, ref } from 'vue'
 import { store } from '@/stores/store'
 import { DynamicScroller, DynamicScrollerItem } from 'vue-virtual-scroller'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
@@ -150,6 +166,7 @@ import ToolTipComponent from './ToolTipComponent.vue'
 import { MetricType, metricToolTips } from '@/model/MetricType'
 import NameElement from './NameElement.vue'
 import { router } from '@/router'
+import ComparisonTableFilter from './ComparisonTableFilter.vue'
 
 library.add(faUserGroup)
 
@@ -161,10 +178,93 @@ const props = defineProps({
   clusters: {
     type: Array<Cluster>,
     required: false
+  },
+  header: {
+    type: String,
+    default: 'Top Comparisons:'
   }
 })
 
-const comparisonList = toRef(props, 'topComparisons')
+const displayedComparisons = computed(() => {
+  const comparisons = getFilteredComparisons(getSortedComparisons(props.topComparisons))
+  let index = 1
+  comparisons.forEach((c) => {
+    c.id = index++
+  })
+  return comparisons
+})
+
+const searchString = ref('')
+
+/**
+ * This function gets called when the search bar for the comparison table has been updated.
+ * It updates the displayed comparisons to only show the ones that  have part of any search result in their id. The search is not case sensitive. The parts can be separated by commas or spaces.
+ * It also updates the anonymous set to unhide a submission if its name was typed in the search bar at any point in time.
+ *
+ * @param newVal The new value of the search bar
+ */
+function getFilteredComparisons(comparisons: ComparisonListElement[]) {
+  const searches = searchString.value
+    .trimEnd()
+    .toLowerCase()
+    .split(/ +/g)
+    .map((s) => s.trim().replace(/,/g, ''))
+  if (searches.length == 0) {
+    return comparisons
+  }
+
+  return comparisons.filter((c) => {
+    const id1 = c.firstSubmissionId.toLowerCase()
+    const id2 = c.secondSubmissionId.toLowerCase()
+    return searches.some((s) => id1.includes(s) || id2.includes(s))
+  })
+}
+
+function getClusterIndexFor(id1: string, id2: string) {
+  let clusterIndex = -1
+  props.clusters?.forEach((c: Cluster, index: number) => {
+    if (c.members.includes(id1) && c.members.includes(id2) && c.members.length > 2) {
+      clusterIndex = index
+    }
+  })
+  return clusterIndex
+}
+
+function getSortedComparisons(comparisons: ComparisonListElement[]) {
+  comparisons.sort(
+    (a, b) =>
+      b.similarities[store().uiState.comparisonTableSortingMetric] -
+      a.similarities[store().uiState.comparisonTableSortingMetric]
+  )
+
+  if (store().uiState.comparisonTableClusterSorting) {
+    comparisons.sort(
+      (a, b) =>
+        getClusterIndexFor(b.firstSubmissionId, b.secondSubmissionId) -
+        getClusterIndexFor(a.firstSubmissionId, a.secondSubmissionId)
+    )
+
+    comparisons.sort(
+      (a, b) =>
+        getClusterFor(b.firstSubmissionId, b.secondSubmissionId).averageSimilarity -
+        getClusterFor(a.firstSubmissionId, a.secondSubmissionId).averageSimilarity
+    )
+  }
+
+  let index = 0
+  comparisons.forEach((c) => {
+    c.sortingPlace = index++
+  })
+  return props.topComparisons
+}
+
+function getClusterFor(id1: string, id2: string) {
+  const index = getClusterIndexFor(id1, id2)
+  if (index < 0 || !props.clusters) {
+    return { averageSimilarity: 0 }
+  }
+  return props.clusters[index]
+}
 
 const displayClusters = props.clusters != undefined
