+++ report-viewer/src/components/SearchBarComponent.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script setup lang="ts">
-import { ref, watch } from 'vue'
+import { computed } from 'vue'
 import Interactable from './InteractableComponent.vue'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
@@ -26,19 +26,30 @@ import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons'
 
 library.add(faMagnifyingGlass)
 
-defineProps({
+const props = defineProps({
   placeholder: {
     type: String,
     default: 'Search...',
     required: false
+  },
+  modelValue: {
+    type: String,
+    default: '',
+    required: false
   }
 })
 
-const emit = defineEmits(['inputChanged', 'searchClicked'])
+const emit = defineEmits<{
+  (e: 'inputChanged', v: string): void
+  (e: 'searchClicked', v: string): void
+  (e: 'update:modelValue', v: string): void
+}>()
 
-const inputText = ref('')
-
-watch(inputText, (newVal) => {
-  emit('inputChanged', newVal)
+const inputText = computed({
+  get: () => props.modelValue,
+  set: (value) => {
+    emit('update:modelValue', value)
+    emit('inputChanged', value)
+  }
 })
 </script>
