+++ report-viewer/src/views/ClusterView.vue
@@ -16,8 +16,11 @@
         <ClusterRadarChart :cluster="clusterListElement" class="flex-grow" />
       </Container>
       <Container class="flex max-h-0 min-h-full w-1/3 flex-col space-y-2">
-        <h2>Comparisons of Cluster Members:</h2>
-        <ComparisonsTable :topComparisons="comparisons" class="min-h-0 flex-1">
+        <ComparisonsTable
+          :topComparisons="comparisons"
+          class="min-h-0 flex-1"
+          header="Comparisons of Cluster Members:"
+        >
           <template #footer v-if="comparisons.length < maxAmountOfComparisonsInCluster">
             <p class="w-full pt-1 text-center font-bold">
               Not all comparisons inside the cluster are shown. To see more, re-run JPlag with a
